backend/requirements.txt
fastapi
uvicorn
pymongo
redis
pika
python-multipart
pydantic-settings
python-dotenv
playwright
pydantic
requests